Refactor bower.json and add Organiser field
Refactor the bower.json file by adding the Organiser field with the value "arazumut". This change improves the organization of the project's metadata.

Translate comments in noop.js to Turkish

Translate the comments in the noop.js file from English to Turkish. This change improves the readability of the code for Turkish-speaking developers.

Translate delta.js constants to Turkish

Translate the constants in the delta.js file from English to Turkish. This change improves the consistency and clarity of the code for Turkish-speaking developers.

Translate comments in filter_visible.js to Turkish

Translate the comments in the filter_visible.js file from English to Turkish. This change improves the readability of the code for Turkish-speaking developers.

Update index.html in devtools/regl_codegen

Update the index.html file in the devtools/regl_codegen directory. Import the Plotly module and remove the unused source script tag. This change ensures that the correct version of Plotly is imported and improves the performance of the code.

Update cross_trace_defaults.js in components/colorscale

Update the cross_trace_defaults.js file in the components/colorscale directory. Add a comment to stash a reference to the color axis. This change improves the maintainability of the code.

Refactor tick_mark_defaults.js in plots/cartesian

Refactor the tick_mark_defaults.js file in the plots/cartesian directory. Update the options comment to indicate that it inherits outerTicks from axes.handleAxisDefaults. This change improves the clarity of the code.

Translate index.js in components/modebar to Turkish

Translate the index.js file in the components/modebar directory from English to Turkish. This change improves the consistency and clarity of the code for Turkish-speaking developers.

Translate label_texttemplate.js comments to Turkish

Translate the comments in the label_texttemplate.js file from English to Turkish. This change improves the readability of the code for Turkish-speaking developers.

Translate index.js in components/sliders to Turkish

Translate the index.js file in the components/sliders directory from English to Turkish. This change improves the consistency and clarity of the code for Turkish-speaking developers.

Translate index.js in components/updatemenus to Turkish

Translate the index.js file in the components/updatemenus directory from English to Turkish. This change improves the consistency and clarity of the code for Turkish-speaking developers.

Refactor calc.js in traces/candlestick

Refactor the calc.js file in the traces/candlestick directory. Update the ptFunc comment to indicate that it is a point function. This change improves the clarity of the code.

Translate mod.js comments to Turkish

Translate the comments in the mod.js file from English to Turkish. This change improves the readability of the code for Turkish-speaking developers.

Refactor layout_attributes.js in components/colorscale

Refactor the layout_attributes.js file in the components/colorscale directory. Remove the unnecessary description comment. This change improves the maintainability of the code.

Translate constants.js in plots/smith to Turkish

Translate the constants.js file in the plots/smith directory from English to Turkish. This change improves the consistency and clarity of the code for Turkish-speaking developers.

Refactor constants.js in components/selections

Refactor the constants.js file in the components/selections directory. Update the comments to improve clarity. This change improves the maintainability of the code.

Refactor constants.js in components/shapes/draw_newshape

Refactor the constants.js file in the components/shapes/draw_newshape directory. Update the comments to improve clarity. This change improves the maintainability of the code.

Refactor constants.js in components/rangeselector

Refactor the constants.js file in the components/rangeselector directory. Update the comments to improve clarity. This change improves the maintainability of the code.